remove duplicate code (#27)
* remove duplicate code


* remove extra reconcile


* change order of handleFinalizer


---------

+++ internal/controller/pullrequest_controller.go
@@ -74,15 +74,26 @@ func (r *PullRequestReconciler) Reconcile(ctx context.Context, req ctrl.Request)
 		return ctrl.Result{}, err
 	}
 
+	deleted, err := r.handleFinalizer(ctx, &pr, pullRequestProvider)
+	if err != nil {
+		return ctrl.Result{}, err
+	}
+	if deleted {
+		return ctrl.Result{}, nil
+	}
+
 	found, err := pullRequestProvider.FindOpen(ctx, &pr)
 	if err != nil {
 		return ctrl.Result{}, err
 	}
 
-	err = r.handleFinalizer(ctx, &pr, pullRequestProvider)
-	if err != nil {
-		return ctrl.Result{}, err
-	}
+	//deleted, err := r.handleFinalizer(ctx, &pr, pullRequestProvider)
+	//if err != nil {
+	//	return ctrl.Result{}, err
+	//}
+	//if deleted {
+	//	return ctrl.Result{}, nil
+	//}
 
 	if !found && pr.Status.State != "" {
 		err := r.Delete(ctx, &pr)
@@ -189,7 +200,7 @@ func (r *PullRequestReconciler) getPullRequestProvider(ctx context.Context, pr p
 	}
 }
 
-func (r *PullRequestReconciler) handleFinalizer(ctx context.Context, pr *promoterv1alpha1.PullRequest, pullRequestProvider scms.PullRequestProvider) error {
+func (r *PullRequestReconciler) handleFinalizer(ctx context.Context, pr *promoterv1alpha1.PullRequest, pullRequestProvider scms.PullRequestProvider) (bool, error) {
 	// name of our custom finalizer
 	finalizerName := "pullrequest.promoter.argoporoj.io/finalizer"
 
@@ -199,7 +210,6 @@ func (r *PullRequestReconciler) handleFinalizer(ctx context.Context, pr *promote
 		// then lets add the finalizer and update the object. This is equivalent
 		// to registering our finalizer.
 		if !controllerutil.ContainsFinalizer(pr, finalizerName) {
-			controllerutil.AddFinalizer(pr, finalizerName)
 			err := retry.RetryOnConflict(retry.DefaultRetry, func() error {
 				err := r.Get(ctx, client.ObjectKeyFromObject(pr), pr)
 				if err != nil {
@@ -209,7 +219,7 @@ func (r *PullRequestReconciler) handleFinalizer(ctx context.Context, pr *promote
 				return r.Update(ctx, pr)
 			})
 			if err != nil {
-				return err
+				return false, err
 			}
 		}
 	} else {
@@ -220,18 +230,19 @@ func (r *PullRequestReconciler) handleFinalizer(ctx context.Context, pr *promote
 			if err != nil {
 				// if fail to delete the external dependency here, return with error
 				// so that it can be retried.
-				return err
+				return false, err
 			}
 
 			// remove our finalizer from the list and update it.
 			controllerutil.RemoveFinalizer(pr, finalizerName)
 			if err := r.Update(ctx, pr); err != nil {
-				return err
+				return true, err
 			}
+			return true, nil
 		}
 	}
 
-	return nil
+	return false, nil
 }
 
 func (r *PullRequestReconciler) createPullRequest(ctx context.Context, pr *promoterv1alpha1.PullRequest, pullRequestProvider scms.PullRequestProvider) error {
